Factor out interaction logic from force vector tool.
